requirements.txt
Django==1.11
django-taggit==0.23.0
django-widget-tweaks==1.4.3
Pillow==5.3.0
pytz==2018.7
